Endstop: overhauled the design to use a TCST1103 optical interrupter.
The design is a bit a compromise. As we want the LED to light up
only with an interrupted switch, there is a balance between
sending voltage to the signal line and sending current to the LED.
R3 has most influence in this balance. Bigger values give more current
to the signal, lower values more current to the LED.

The current balance gives 4 Volts to the signal and lets the LED
light up to about 80%.

Another design solution would be to connect the LED to Vcc instead of
GND, giving both, full voltage to the signal and full current to the LED.
However, the signal is inverted then, lighting the LED all the time.
Adding a transistor, too, would get the LED lighting right, but also
make the circuit more complex.